adding builder dockerfile

+++ src/main/java/com/sepro/Authenticationserver/entity/BaseIdEntity.java
@@ -8,5 +8,5 @@ public class BaseIdEntity {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "ID")
-    protected int id;
+    protected Long id;
 }
